refactor: 워크스페이스 패널 분리 및 lazy loading 적용
